+++ js/directorylisting.js
@@ -1,12 +1,19 @@
 $(document).ready(function() {
+    // Open first level by default
 	$('div.directorylist > ul.directorylisting > li.item.directory > input[type="checkbox"]').prop('checked', true);
 	$('div.directorylist > ul.directorylisting > ul.directorylisting').show();
-	
-	$('div.directorylist li.item.directory input[type="checkbox"]').change(function() {
-		if ($(this).is(":checked")) {
-			$(this).parent().next("ul.directorylisting").show();
-		} else {
-			$(this).parent().next("ul.directorylisting").hide();
-		}
-	});
+    
+    // Iterate each Directory List and open folders on check
+    $("div.directorylist").each( function( index, element ) {
+        $('li.item.directory input[type="checkbox"]', this).each( function( index, element ) {
+            $(this).change(function() {
+                if ($(this).is(":checked")) {
+                    $('li.item.directory').next("ul.directorylisting").show();
+                } else {
+                    $('li.item.directory').next("ul.directorylisting").hide();
+                }
+            });
+        });
+    });
+
 });